Remove creating deb/rpm packages by goreleaser. Update setup docs.
Creating deb/rpm packages temporarily removed (may be returned in future). The main reason is usecases used in package-way is different than in targz-way. Auto-update mode becomes not working when pgscv installed from packages. Thus we need to re-organize packages structure and implement the same setup as in the built-in bootstrap.